Changed all simulation events to use int for time instead of long

+++ src/v2/core/events/AbstractSimulationEvent.java
@@ -5,9 +5,9 @@ package v2.core.events;
  */
 public class AbstractSimulationEvent implements SimulationEvent {
 
-    private final long time;
+    private final int time;
 
-    public AbstractSimulationEvent(long time) {
+    public AbstractSimulationEvent(int time) {
         this.time = time;
     }
